mise a jour du projet

+++ src/main/java/com/example/aspose_backend/repository/ModeleRepository.java
@@ -3,5 +3,4 @@ package com.example.aspose_backend.repository;
 import com.example.aspose_backend.model.Modele;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ModeleRepository extends JpaRepository<Modele, Integer> {
-}
+public interface ModeleRepository extends JpaRepository<Modele, Long> {}
